Committed to Fortnite

+++ src/main/java/org/Game/CollisionChecker.java
@@ -25,7 +25,7 @@ public class CollisionChecker {
 
         switch(entity.direction){
             case "up":
-                entityTopRow = (entityTopWorldY - entity.speed)/gP.tileSize;
+                entityTopRow = (entityTopWorldY - (int)entity.speed)/gP.tileSize;
                 tileNum1 = gP.tileM.mapTileNum[entityLeftCol][entityTopRow];
                 tileNum2 = gP.tileM.mapTileNum[entityRightCol][entityTopRow];
                 if (gP.tileM.tile[tileNum1].collision ||
@@ -34,7 +34,7 @@ public class CollisionChecker {
                 }
                 break;
             case "down":
-                entityBottomRow = (entityBottomWorldY + entity.speed)/gP.tileSize;
+                entityBottomRow = (entityBottomWorldY + (int)entity.speed)/gP.tileSize;
                 tileNum1 = gP.tileM.mapTileNum[entityLeftCol][entityBottomRow];
                 tileNum2 = gP.tileM.mapTileNum[entityRightCol][entityBottomRow];
                 if (gP.tileM.tile[tileNum1].collision ||
@@ -43,7 +43,7 @@ public class CollisionChecker {
                 }
                 break;
             case "left":
-                entityLeftCol = (entityLeftWorldX - entity.speed)/gP.tileSize;
+                entityLeftCol = (entityLeftWorldX - (int)entity.speed)/gP.tileSize;
                 tileNum1 = gP.tileM.mapTileNum[entityLeftCol][entityTopRow];
                 tileNum2 = gP.tileM.mapTileNum[entityLeftCol][entityBottomRow];
                 if (gP.tileM.tile[tileNum1].collision ||
@@ -52,7 +52,7 @@ public class CollisionChecker {
                 }
                 break;
             case "right":
-                entityRightCol = (entityRightWorldX + entity.speed)/gP.tileSize;
+                entityRightCol = (entityRightWorldX + (int)entity.speed)/gP.tileSize;
                 tileNum1 = gP.tileM.mapTileNum[entityRightCol][entityTopRow];
                 tileNum2 = gP.tileM.mapTileNum[entityRightCol][entityBottomRow];
                 if (gP.tileM.tile[tileNum1].collision ||
